// src/swagger/model/updateVehicleDetailsDto.ts
/**
 * fms
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface UpdateVehicleDetailsDto { 
    registration_No?: string;
    vehicleSubStatus?: string;
    standBy_FromDate?: string;
    standBy_ToDate?: string;
    onRoute_FromDate?: string;
    onRoute_ToDate?: string;
    off_Road?: string;
    offroad_Date?: string;
    offroad_ToDate?: string;
    offroad_Reason?: string;
}